fix: eagerly declare bindings for variables in scope
Motivation: Consider this code snippet:

```
var a = 3;
{
  fun g() {
    print a;
  }
  g(); // (1)
  var a = 4;
  g(); // (2)
}
```

Intuitively, `a` in `g` should either bind to the global `a` or the
local `a`. However, currently, a declaration of `a` in the local scope
changes the binding target of `a`. This commit mandates that the closest
variable in scope is always bound to, even though it's not immediately
visible. In the code snippet above, (1) would throw an error that says
"Variable 'a' used before defined".

This commit does this by scanning the scope for declarations and
creating a dummy binding for each.